MSM8996 AUTO: Fix the webview provider issue for Non GMS build.
In OMR1, it is not picking the android default WebView in the non GMS build.
Because config_webview_packages.xml file for GoogleWebview inside Car Product
is overriding android default WebView with the GoogleWebView.

Fix: config_webview_packages.xml inside Car product is not required
because it is already present inside partner_GMS for GoogleWebview.
So removed the file.

Tested: Ran CTSWebViewTestCase in non GMS build. All the test cases
passed.

Change-Id: Ib5e0beb51acbabc21b4a96507b93b03764685b09