Car product: Set device provisioned by default
In HQX, device is not provisioned by default because of missing
setup wizard and some apps which prevents some functionalities
from working properly. Setting this manually over here for testing
purposes. However, customer has no need of this since they would
be using the setup wizard to setup the user.

Change-Id: I90308c23c31dfbb27e99dbbb40b504857566533d