CarService: Optimize LPM sleep time in CPMS
1. Add the property "android.car.enable_cpms_polling"
to enable/disable CPMS polling dynamically.

(1) The default value is true, indicating CPMS start
the timer for ShutdownProcessingTimerTask when to
suspend, before it calls the API to force suspend
(i.e. $ echo mem > /sys/power/state).

By default, CPMS poll every 2 seconds to wait for app's
completion in 15 minutes.

(2) If the property is set as false, the timer for
ShutdownProcessingTimerTask isn't started up.

$ adb shell setprop android.car.enable_cpms_polling false

2. Stop listening Car power state in GarageModeService,
if Garage Mode is stopped.

$ adb shell dumpsys car_service garage-mode off

Note:
(1) If Garage Mode is enabled, CPMS take 10 seconds
at least to wait for Garage Mode's completion when
to suspend.

Change-Id: I9fde363394a0911ca96bf44af6f880af94552dbe